examples/demo_complex_app/requirements.txt
Flask==2.3.3
requests>=2.31.0
openai>=1.0.0
PyJWT>=2.8.0
schedule>=1.2.0